Project Framework?
Okay, I added the basic framework that I borrowed from a previous game that I made.  This can be edited as needed, and if anyone has questions about whats going on, I can go over it sometime in class.  I know it sucks for you all since you weren't coding along for this part so it may be confusing at first, but even if it takes a week or two for me to explain and to move past it, it will still be a pretty big time save to use this framework.  Hopefully the upload works correctly.  I use IntelliJ so if there is a weird formatting or transfer issue just let me know.  Feel free to check out the code and even start messing around with it.  Have a great weekend!